Implement more tests
More tests for creating events have been added. These are successfully
creating an event and checking that the venue can't be null. To
implement this some changes had to be made as before venue objects where
being passed from the form, however this could not be mimicked in tests.
Therefore, venue id has been used instead. This now has an attribute in
event objects and is passed from the form, and then in the controller
the venue object for an event is set using this id. Additionally a
previous failing test now passes.

Issue: #14