Fix problem with updating events
Changed functionality of the update event page to always select the
venue of the event it already has. Before that change it was always
preselecting the first option no matter the actual value of the venue
field. Additionally, fixed showing events page to handle null
descriptions and empty string descriptions.